requirements.txt
httpx>=0.27.0
mcp[cli]>=1.26.0
python-dotenv>=1.0.0
